ref(related_issues): TraceTimelineOrRelatedIssue component
This helps testing the TraceLink, TraceTimeline and TraceIssueEvent as a collective of related component.

Before, we could not test the behaviour of the TraceLink.
