+++ static/app/views/issueDetails/traceTimelineOrRelatedIssue.spec.tsx
@@ -8,13 +8,14 @@ import ProjectsStore from 'sentry/stores/projectsStore';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import useRouteAnalyticsParams from 'sentry/utils/routeAnalytics/useRouteAnalyticsParams';
 
-import {TraceTimeline} from './traceTimeline';
-import type {TraceEventResponse} from './useTraceTimelineEvents';
+import {TraceTimeline} from './traceTimeline/traceTimeline';
+import type {TraceEventResponse} from './traceTimeline/useTraceTimelineEvents';
+import {TraceTimeLineOrRelatedIssue} from './traceTimelineOrRelatedIssue';
 
 jest.mock('sentry/utils/routeAnalytics/useRouteAnalyticsParams');
 jest.mock('sentry/utils/analytics');
 
-describe('TraceTimeline', () => {
+describe('TraceTimeline & TraceRelated Issue', () => {
   // Paid plans have global-views enabled
   // Include project: -1 in all matchQuery calls to ensure we are looking at all projects
   const organization = OrganizationFixture({
@@ -100,13 +101,14 @@ describe('TraceTimeline', () => {
       body: twoIssuesBody,
       match: [MockApiClient.matchQuery({dataset: 'discover', project: -1})],
     });
-    render(<TraceTimeline event={event} />, {organization});
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
+      organization,
+    });
     expect(await screen.findByLabelText('Current Event')).toBeInTheDocument();
 
     await userEvent.hover(screen.getByTestId('trace-timeline-tooltip-1'));
     expect(await screen.findByText('You are here')).toBeInTheDocument();
     expect(useRouteAnalyticsParams).toHaveBeenCalledWith({
-      has_related_trace_issue: false,
       trace_timeline_status: 'shown',
     });
   });
@@ -122,11 +124,12 @@ describe('TraceTimeline', () => {
       body: discoverBody,
       match: [MockApiClient.matchQuery({dataset: 'discover', project: -1})],
     });
-    const {container} = render(<TraceTimeline event={event} />, {organization});
+    const {container} = render(<TraceTimeline event={event} />, {
+      organization,
+    });
     await waitFor(() =>
       expect(useRouteAnalyticsParams).toHaveBeenCalledWith({
-        has_related_trace_issue: false,
-        trace_timeline_status: 'empty',
+        trace_timeline_status: 'shown',
       })
     );
     expect(container).toBeEmptyDOMElement();
@@ -143,10 +146,11 @@ describe('TraceTimeline', () => {
       body: emptyBody,
       match: [MockApiClient.matchQuery({dataset: 'discover', project: -1})],
     });
-    const {container} = render(<TraceTimeline event={event} />, {organization});
+    const {container} = render(<TraceTimeline event={event} />, {
+      organization,
+    });
     await waitFor(() =>
       expect(useRouteAnalyticsParams).toHaveBeenCalledWith({
-        has_related_trace_issue: false,
         trace_timeline_status: 'empty',
       })
     );
@@ -164,7 +168,9 @@ describe('TraceTimeline', () => {
       body: twoIssuesBody,
       match: [MockApiClient.matchQuery({dataset: 'discover', project: -1})],
     });
-    render(<TraceTimeline event={event} />, {organization});
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
+      organization,
+    });
     // Checking for the presence of seconds
     expect(await screen.findAllByText(/\d{1,2}:\d{2}:\d{2} (AM|PM)/)).toHaveLength(5);
   });
@@ -184,7 +190,9 @@ describe('TraceTimeline', () => {
       },
       match: [MockApiClient.matchQuery({dataset: 'discover', project: -1})],
     });
-    render(<TraceTimeline event={event} />, {organization});
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
+      organization,
+    });
     expect(await screen.findByLabelText('Current Event')).toBeInTheDocument();
   });
 
@@ -205,7 +213,7 @@ describe('TraceTimeline', () => {
       body: [],
     });
 
-    render(<TraceTimeline event={event} />, {
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
       organization,
     });
 
@@ -220,7 +228,6 @@ describe('TraceTimeline', () => {
     await userEvent.click(await screen.findByText('Slow DB Query'));
     expect(useRouteAnalyticsParams).toHaveBeenCalledWith({
       has_related_trace_issue: true,
-      trace_timeline_status: 'empty',
     });
     expect(trackAnalytics).toHaveBeenCalledTimes(1);
     expect(trackAnalytics).toHaveBeenCalledWith(
@@ -250,7 +257,7 @@ describe('TraceTimeline', () => {
       body: [],
     });
 
-    render(<TraceTimeline event={event} />, {
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
       organization,
     });
 
@@ -262,10 +269,7 @@ describe('TraceTimeline', () => {
 
     // We do not display the timeline because we only have 1 event
     expect(await screen.queryByLabelText('Current Event')).not.toBeInTheDocument();
-    expect(useRouteAnalyticsParams).toHaveBeenCalledWith({
-      has_related_trace_issue: false,
-      trace_timeline_status: 'empty',
-    });
+    expect(useRouteAnalyticsParams).toHaveBeenCalledTimes(0);
   });
 
   it('trace timeline works for plans with no global-views feature', async () => {
@@ -292,7 +296,7 @@ describe('TraceTimeline', () => {
       ],
     });
 
-    render(<TraceTimeline event={event} />, {
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
       organization: OrganizationFixture({
         features: [],
       }),
@@ -329,7 +333,7 @@ describe('TraceTimeline', () => {
       body: [],
     });
 
-    render(<TraceTimeline event={event} />, {
+    render(<TraceTimeLineOrRelatedIssue event={event} />, {
       organization: OrganizationFixture({
         features: [],
       }),
